+++ src/apps/seelenweg/modules/shared/utils/domain.ts
@@ -1,0 +1,20 @@
+export const UWP_TARGET_SIZE_POSTFIXS = [
+  '.targetsize-256_altform-lightunplated.png',
+  '.targetsize-256.png',
+  '.targetsize-96_altform-lightunplated.png',
+  '.targetsize-96.png',
+  '.targetsize-64_altform-lightunplated.png',
+  '.targetsize-64.png',
+  '.targetsize-48_altform-lightunplated.png',
+  '.targetsize-48.png',
+  '.targetsize-32_altform-lightunplated.png',
+  '.targetsize-32.png',
+];
+
+export const UWP_SCALE_POSTFIXS = [
+  '.scale-400.png',
+  '.scale-200.png',
+  '.scale-150.png',
+  '.scale-125.png',
+  '.scale-100.png',
+];