+++ src/background/utils/mod.rs
@@ -1,0 +1,4 @@
+
+pub fn filename_from_path(path: &str) -> String {
+    path.split('\\').last().unwrap_or_default().to_string()
+}